Refactor: JpaRepository를 Repository 인터페이스로 합치고  repository 패키지로 이동함
repository를 하나로 합치고, 그에 따른 다른 클래스들 import문 변경됨
