+++ app/src/main/java/com/codesoom/assignment/repository/ProductRepository.java
@@ -1,14 +1,14 @@
-package com.codesoom.assignment.infra;
+package com.codesoom.assignment.repository;
 
 import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.domain.ProductRepository;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface JpaProductRepository
-        extends ProductRepository, CrudRepository<Product, Long> {
+@Repository
+public interface ProductRepository extends CrudRepository<Product, Long> {
     List<Product> findAll();
 
     Optional<Product> findById(Long id);
