+++ app/src/test/java/com/codesoom/assignment/application/ProductServiceTest.java
@@ -8,7 +8,7 @@ package com.codesoom.assignment.application;
 
 import com.codesoom.assignment.errors.ProductNotFoundException;
 import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.domain.ProductRepository;
+import com.codesoom.assignment.repository.ProductRepository;
 import com.codesoom.assignment.dto.ProductData;
 import com.github.dozermapper.core.DozerBeanMapperBuilder;
 import com.github.dozermapper.core.Mapper;
